feat: add exploit class methods into cvedb class
