+++ cve_bin_tool/cli.py
@@ -41,7 +41,6 @@ from cve_bin_tool.error_handler import (
     InsufficientArgs,
     excepthook,
 )
-from cve_bin_tool.exploitdb import ExploitChecker
 from cve_bin_tool.input_engine import InputEngine, TriageData
 from cve_bin_tool.log import LOGGER
 from cve_bin_tool.merge import MergeReports
@@ -419,9 +418,8 @@ def main(argv=None):
 
     # Connect to and update exploits database
     if args["update_exploits"] and not args["offline"]:
-        exploits = ExploitChecker()
-        exploits.create_exploit_db()
-        exploits.update_exploits()
+        cvedb_orig.create_exploit_db()
+        cvedb_orig.update_exploits()
 
     # if OLD_CACHE_DIR (from cvedb.py) exists, print warning
     if os.path.exists(OLD_CACHE_DIR):
@@ -507,7 +505,13 @@ def main(argv=None):
             )
         )
 
-    with CVEScanner(score=score, check_exploits=args["exploits"]) as cve_scanner:
+    if args["exploits"]:
+        cvedb_orig.get_cache_exploits()
+    with CVEScanner(
+        score=score,
+        check_exploits=args["exploits"],
+        exploits_list=cvedb_orig.get_exploits_list(),
+    ) as cve_scanner:
         triage_data: TriageData
         total_files: int = 0
         parsed_data: Dict[ProductInfo, TriageData] = {}
