+++ cve_bin_tool/cve_scanner.py
@@ -15,7 +15,6 @@ from rich.console import Console
 
 from cve_bin_tool.cvedb import DBNAME, DISK_LOCATION_DEFAULT
 from cve_bin_tool.error_handler import ErrorMode
-from cve_bin_tool.exploitdb import ExploitChecker
 from cve_bin_tool.input_engine import TriageData
 from cve_bin_tool.log import LOGGER
 from cve_bin_tool.theme import cve_theme
@@ -43,6 +42,7 @@ class CVEScanner:
         logger: Logger = None,
         error_mode: ErrorMode = ErrorMode.TruncTrace,
         check_exploits: bool = False,
+        exploits_list: List[str] = [],
     ):
         self.logger = logger or LOGGER.getChild(self.__class__.__name__)
         self.error_mode = error_mode
@@ -52,6 +52,7 @@ class CVEScanner:
         self.all_cve_data = defaultdict(CVEData)
         self.all_cve_version_info = dict()
         self.check_exploits = check_exploits
+        self.exploits_list = exploits_list
 
     def get_cves(self, product_info: ProductInfo, triage_data: TriageData):
         """Get CVEs against a specific version of a product.
@@ -196,11 +197,6 @@ class CVEScanner:
                 result = self.cursor.execute(query, cve_list[start:end] + [self.score])
                 start = end
 
-                # Checking for exploited CVEs
-                if self.check_exploits:
-                    exploits = ExploitChecker()
-                    exploits.get_cache_exploits()
-
                 for row in result:
                     triage = triage_data.get(row["cve_number"]) or triage_data.get(
                         "default"
@@ -213,9 +209,12 @@ class CVEScanner:
                         row_dict.update(triage)
                         # print(row_dict)
                         row_dict["severity"] = row_dict["severity"] or row["severity"]
-                        if self.check_exploits:
-                            if exploits.check_exploit(row_dict["cve_number"]):
-                                row_dict["severity"] += "-EXPLOIT"
+                        # Checking for exploits
+                        if (
+                            self.check_exploits
+                            and row_dict["cve_number"] in self.exploits_list
+                        ):
+                            row_dict["severity"] += "-EXPLOIT"
                         row_dict["score"] = row_dict["score"] or row["score"]
                         row_dict["cvss_version"] = (
                             row_dict["cvss_version"] or row["cvss_version"]
