+++ cve_bin_tool/cvedb.py
@@ -17,6 +17,7 @@ import shutil
 import sqlite3
 
 import aiohttp
+import requests
 from bs4 import BeautifulSoup
 from rich.progress import track
 
@@ -92,6 +93,8 @@ class CVEDB:
         self.nvd_type = nvd_type
         self.incremental_update = incremental_update
         self.all_cve_entries = None
+        self.exploits_list = []
+        self.exploit_count = 0
 
         if not os.path.exists(self.dbpath):
             self.rollback_cache_backup()
@@ -798,3 +801,62 @@ class CVEDB:
 
     def __del__(self):
         self.rollback_cache_backup()
+
+    # Methods to check and update exploits
+
+    def update_exploits(self):
+        url = "https://www.cisa.gov/sites/default/files/feeds/known_exploited_vulnerabilities.json"
+        r = requests.get(url)
+        data = r.json()
+        cves = data["vulnerabilities"]
+        exploit_list = []
+        for cve in cves:
+            exploit_list.append((cve["cveID"], cve["product"], cve["shortDescription"]))
+        self.populate_exploit_db(exploit_list)
+
+    def get_cache_exploits(self):
+        get_exploits = """
+        SELECT cve_number FROM cve_exploited
+        """
+        self.db_open()
+        cursor = self.connection.cursor()
+        cursor.row_factory = lambda cursor, row: row[0]
+        self.exploits_list = cursor.execute(get_exploits).fetchall()
+        self.db_close()
+        self.exploit_count = len(self.exploits_list)
+
+    def get_exploits_list(self):
+        return self.exploits_list
+
+    def get_exploits_count(self):
+        return self.exploit_count
+
+    def create_exploit_db(self):
+        create_exploit_table = """
+        CREATE TABLE IF NOT EXISTS cve_exploited (
+            cve_number TEXT,
+            product TEXT,
+            description TEXT,
+            PRIMARY KEY(cve_number)
+        )
+        """
+        self.db_open()
+        cursor = self.connection.cursor()
+        cursor.execute(create_exploit_table)
+        self.connection.commit()
+        self.db_close()
+
+    def populate_exploit_db(self, exploits):
+        insert_exploit = """
+        INSERT or REPLACE INTO cve_exploited (
+            cve_number,
+            product,
+            description
+        )
+        VALUES (?,?,?)
+        """
+        self.db_open()
+        cursor = self.connection.cursor()
+        cursor.executemany(insert_exploit, exploits)
+        self.connection.commit()
+        self.db_close()
