--- cve_bin_tool/exploitdb.py
@@ -1,83 +1,0 @@
-# Copyright (C) 2022 Intel Corporation
-# SPDX-License-Identifier: GPL-3.0-or-later
-
-import os
-import sqlite3
-
-import requests
-
-
-class ExploitChecker:
-
-    DBPATH = os.path.join(os.path.expanduser("~"), ".cache", "cve-bin-tool", "cve.db")
-
-    def __init__(self):
-        self.exploits = []
-        self.connection = None
-        self.exploit_count = 0
-
-    def open_connection_db(self):
-        if self.connection is None:
-            self.connection = sqlite3.connect(self.DBPATH)
-
-    def close_connection_db(self):
-        if self.connection:
-            self.connection.close()
-            self.connection = None
-
-    def update_exploits(self):
-        url = "https://www.cisa.gov/sites/default/files/feeds/known_exploited_vulnerabilities.json"
-        r = requests.get(url)
-        data = r.json()
-        cves = data["vulnerabilities"]
-        exploit_list = []
-        for cve in cves:
-            exploit_list.append((cve["cveID"], cve["product"], cve["shortDescription"]))
-        self.populate_exploit_db(exploit_list)
-
-    def get_cache_exploits(self):
-        get_exploits = """
-        SELECT cve_number FROM cve_exploited
-        """
-        self.open_connection_db()
-        cursor = self.connection.cursor()
-        cursor.row_factory = lambda cursor, row: row[0]
-        self.exploits = cursor.execute(get_exploits).fetchall()
-        self.close_connection_db()
-        self.exploit_count = len(self.exploits)
-
-    def create_exploit_db(self):
-        create_exploit_table = """
-        CREATE TABLE IF NOT EXISTS cve_exploited (
-            cve_number TEXT,
-            product TEXT,
-            description TEXT,
-            PRIMARY KEY(cve_number)
-        )
-        """
-        self.open_connection_db()
-        cursor = self.connection.cursor()
-        cursor.execute(create_exploit_table)
-        self.connection.commit()
-        self.close_connection_db()
-
-    def populate_exploit_db(self, exploits):
-        insert_exploit = """
-        INSERT or REPLACE INTO cve_exploited (
-            cve_number,
-            product,
-            description
-        )
-        VALUES (?,?,?)
-        """
-        self.open_connection_db()
-        cursor = self.connection.cursor()
-        cursor.executemany(insert_exploit, exploits)
-        self.connection.commit()
-        self.close_connection_db()
-
-    def check_exploit(self, exploit):
-        return exploit in self.exploits
-
-    def get_exploit_count(self):
-        return self.exploit_count
